Move recaptcha stubs out to common helper
